+++ contracts/TokenLedger.sol
@@ -4,22 +4,22 @@ import "@openzeppelin/contracts/access/Ownable.sol";
 import "../interfaces/ITokenLedger.sol";
 
 contract TokenLedger is Ownable, ITokenLedger {
-    uint16 private sourceChainId;
-    mapping(address => mapping(uint16 => address)) private sourceTokenToTargetToken;
-    mapping(address => mapping(uint16 => address)) private targetTokenToSourceToken;
+    uint32 private sourceChainId;
+    mapping(address => mapping(uint32 => address)) private sourceTokenToTargetToken;
+    mapping(address => mapping(uint32 => address)) private targetTokenToSourceToken;
 
     event TokenConnectionRegistered(
         address indexed sourceToken,
         address indexed targetToken,
-        uint16 sourceChainId,
-        uint16 targetChainId
+        uint32 sourceChainId,
+        uint32 targetChainId
     );
 
     constructor() {
-        sourceChainId = uint16(block.chainid);
+        sourceChainId = uint32(block.chainid);
     }
 
-    function getTargetTokenAddress(address _sourceToken, uint16 _targetChainId)
+    function getTargetTokenAddress(address _sourceToken, uint32 _targetChainId)
         external
         view
         override
@@ -28,7 +28,7 @@ contract TokenLedger is Ownable, ITokenLedger {
         return sourceTokenToTargetToken[_sourceToken][_targetChainId];
     }
 
-    function getSourceTokenAddress(address _targetToken, uint16 _sourceChainId)
+    function getSourceTokenAddress(address _targetToken, uint32 _sourceChainId)
         external
         view
         override
@@ -39,7 +39,7 @@ contract TokenLedger is Ownable, ITokenLedger {
 
     function registerTargetTokenAddress(
         address _sourceToken,
-        uint16 _targetChainId,
+        uint32 _targetChainId,
         address _targetToken
     ) external override {
         require(_sourceToken != address(0), "Invalid source address");
